+++ PythonAPI/test/unit/test_transform.py
@@ -205,6 +205,19 @@ class TestTransform(unittest.TestCase):
         self.assertTrue(abs(t.offset_y - 2.0) <= error)
         self.assertTrue(abs(t.offset_z - 3.0) <= error)
         
+    def test_geo_offset_transform_zero(self):
+        error = 0.001
+        t = carla.GeoOffsetTransform(0.0, 0.0, 0.0, 0.0)
+        loc = carla.Location(1.0, 2.0, 3.0)
+
+        out = t.ApplyTransformation(loc)
+
+        solution_list = carla.Location(1.0, 2.0, 3.0)
+
+        self.assertTrue(abs(out.x - solution_list.x) <= error)
+        self.assertTrue(abs(out.y - solution_list.y) <= error)
+        self.assertTrue(abs(out.z - solution_list.z) <= error)
+
     def test_geo_offset_transform_translation(self):
         error = 0.001
         t = carla.GeoOffsetTransform(10.0, 20.0, 5.0, 0.0)
@@ -212,7 +225,7 @@ class TestTransform(unittest.TestCase):
 
         out = t.ApplyTransformation(loc)
 
-        solution_list = carla.Location(11.0, 18.0, 8.0)
+        solution_list = carla.Location(11.0, -18.0, 8.0)
 
         self.assertTrue(abs(out.x - solution_list.x) <= error)
         self.assertTrue(abs(out.y - solution_list.y) <= error)